Update status checker to verify homepage content
Oops, we got blocked by a new anti-DDoS layer! But our status checker didn't realize it, because it just checks that the network request to http://www.neopets.com/ succeeds at all.

In this change, we add extra logic to the status checker, to also read the page body. We figure that, if the page contains "login" or something similar, it's probably the real homepage; but if it doesn't, we act as if the request failed.

# Test Plan:
1. Run the Neopia server locally, with default settings (i.e. against real Neopets.com). Confirm that the console shows `Neopets.com status: false. Body exists, but didn't contain "login".`, and `/api/1/status` returns false. (After Neopets unblocks us though, this test case should yield true instead!)
2. Visit http://www.neopets.com/ in a logged-out browser, download the HTML content, and serve it on a local web server on port 9000 (or whatever). Then, run the Neopia server locally, with `-pingUrl http://localhost:9000/`. Confirm that the console and `/api/1/status` both show status true.
3. Run the Neopia server locally, with `-pingUrl http://localhost:78213` (or any port you don't have configured for HTTP responses). Confirm that the console shows `Neopets.com status: false. Network request failed.`, and `/api/1/status` returns false.